+++ AddOns/NovaWorldBuffs/Locale/frFR.lua
@@ -157,12 +157,12 @@ L["Options"] = "Options";
 
 --Spirit of Zandalar buff NPC first yell string (part of his first yell msg before before buff).
 L["Begin the ritual"] = "Commencez le rituel"
-L["The Blood God"] = "Le Dieu sanglant"; --First Booty bay yell from Zandalarian Emissary.
+L["The Blood God"] = "Le Dieu sanglantd"; --First Booty bay yell from Zandalarian Emissary.
 --Spirit of Zandalar buff NPC second yell string (part of his second yell msg before before buff).
 L["slayer of Hakkar"] = "vainqueur d’Hakkar";
 
 L["Molthor"] = "Molthor";
-L["Emissaire zandalarien"] = "Zandalarian Emissary";
+L["Zandalarian Emissary"] = "Emissaire zandalarien";
 L["Whipper Root Tuber"] = "Tubercule de navetille";
 L["Night Dragon's Breath"] = "Souffle de dragon nocturne";
 L["Resist Fire"] = "Résistance au Feu"; -- LBRS fire resist buff.