+++ AddOns/Prat-3.0/addon/addon.lua
@@ -76,7 +76,7 @@ Version = "Prat |cff8080ff3.0|r (|cff8080ff" .. "DEBUG" .. "|r)"
 --@end-debug@]===]
 
 --@non-debug@
-Version = "Prat |cff8080ff3.0|r (|cff8080ff".."3.7.65".."|r)"
+Version = "Prat |cff8080ff3.0|r (|cff8080ff".."3.7.66".."|r)"
 --@end-non-debug@