+++ AddOns/Prat-3.0/modules/Timestamps.lua
@@ -24,7 +24,12 @@
 --
 -------------------------------------------------------------------------------
 Prat:AddModuleToLoad(function()
+  local function dbg(...) end
 
+  --[===[@debug@
+  function dbg(...) Prat:PrintLiteral(...) end
+
+  --@end-debug@]===]
 
   local PRAT_MODULE = Prat:RequestModuleName("Timestamps")
 
@@ -571,12 +576,9 @@ L = {
   end)
 
   function module:OnModuleEnable()
-    -- For this module to work, it must hook before Prat
     for _, v in pairs(Prat.HookedFrames) do
-      self:RawHook(v, "AddMessage", true)
+      self:SecureHook(v, "AddMessage")
     end
-
-    self:RawHook("ChatChannelDropDown_PopOutChat", true)
 
     self.secondsDifference = 0
     self.lastMinute = select(2, GetGameTime())
@@ -594,7 +596,7 @@ L = {
   function module:Prat_FramesUpdated(info, name, chatFrame, ...)
     if not hookedFrames[chatFrame:GetName()] then
       hookedFrames[chatFrame:GetName()] = true
-      self:RawHook(chatFrame, "AddMessage", true)
+      self:SecureHook(chatFrame, "AddMessage")
     end
   end
 
@@ -605,20 +607,14 @@ L = {
     end
   end
 
-  function module:ChatChannelDropDown_PopOutChat(...)
-    Prat.loading = true
-    self.hooks["ChatChannelDropDown_PopOutChat"](...)
-    Prat.loading = nil
-  end
-
   --[[------------------------------------------------
       Core Functions
   ------------------------------------------------]] --
   function module:AddMessage(frame, text, ...)
-    if self.db.profile.show and self.db.profile.show[frame:GetName()] and not Prat.loading then
-      text = text and self:InsertTimeStamp(text, frame)
+    if self.db.profile.show and self.db.profile.show[frame:GetName()] then
+      frame:TransformMessages(function(message, ...) return message == text end,
+            function(message, ...) return self:InsertTimeStamp(message, frame), ... end)
     end
-    self.hooks[frame].AddMessage(frame, text, ...)
   end
 
   function module:IsTimestampPlain()