+++ AddOns/Prat-3.0/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Prat 3.0
 
-## [3.7.65](https://github.com/sylvanaar/prat-3-0/tree/3.7.65) (2020-05-17)
-[Full Changelog](https://github.com/sylvanaar/prat-3-0/compare/3.7.64...3.7.65)
+## [3.7.66](https://github.com/sylvanaar/prat-3-0/tree/3.7.66) (2020-05-19)
+[Full Changelog](https://github.com/sylvanaar/prat-3-0/compare/3.7.65...3.7.66)
 
-- ChatFrames: Fix the clamping of the combat log which is set at a different time than the chatframes  
+- ScrollBack: Use Alternative API to load the chatframe scrollback so it doesn't get a second timestamp  
+- Paragraph: Support Setting The ChatFrame Line Spacing  
+- Timestamps: Prototype "Taint-Free" Implementation (#59). Only uses SecureHook  
+- History: Don't save secure chat commands in the tainted history